Combined Stop/PState Cme Image + unified interrupt handler
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/20857
Tested-by: Jenkins Server

+++ import/chips/p9/procedures/ppe_closed/sgpe/stop_gpe/p9_hcd_cache_ras_runtime_scom.c
@@ -27,7 +27,7 @@
 #include "p9_sgpe_stop_exit_marks.h"
 
 int
-p9_hcd_cache_ras_runtime_scom(uint8_t quad)
+p9_hcd_cache_ras_runtime_scom(uint32_t quad)
 {
     return 0;
 }
