+++ import/chips/p9/procedures/ppe_closed/sgpe/stop_gpe/p9_hcd_cache_scomcust.c
@@ -27,7 +27,7 @@
 #include "p9_sgpe_stop_exit_marks.h"
 
 int
-p9_hcd_cache_scomcust(uint8_t quad)
+p9_hcd_cache_scomcust(uint32_t quad)
 {
     int rc = SGPE_STOP_SUCCESS;
